% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_align_me.R
\name{align_me}
\alias{align_me}
\title{Align time-course data based on an Mixed-Effects alignment model}
\usage{
align_me(
  data,
  model = "yi / sj",
  errmodel = "value * sigmaR",
  fixed = ys ~ condition,
  scaling = sj ~ replicate,
  error = sigmaR ~ 1,
  log = TRUE,
  normalize = TRUE,
  reduce = FALSE,
  verbose = FALSE,
  normalize_input = TRUE
)
}
\arguments{
\item{data}{data.frame containing the the data to be scaled. Usualy the
output of \link{read_wide} is used. Obligatory are the columns \code{name},
and \code{value}. In the case of time dependet data, a \code{time} column is
also necessary. Additionally, \code{data} should have further columns,
e.g. characterizing experimental conditions (the fixed effects) and
sources of variance between data of the same experimental condition
(the scaled variables).}

\item{model}{character defining the model by which the values in
\code{data} can be described, e.g. "yi/sj"}

\item{errmodel}{character defining a model for the standard
deviation of a value, e.g. "sigma0 + value * sigmaR". This model
can contain parameters, e.g. "sigma0" and "sigmaR", or numeric
variables from \code{data}, e.g. "value" or "time".}

\item{fixed}{two-sided formula of the form
\code{par1+par2+... ~ name1+name2+...} where "par1, par2, ..." are
parameters contained in \code{model}, e.g. "yi", and "name1, ..."
refers to variables in \code{data}, e.g. "condition".
The parameters "par1, ..." are determined specific to the levels of
"name1, ...".}

\item{scaling}{two-sided formula of the form
\code{par1+par2+... ~ name1+name2+...} where "par1, par2, ..." are
parameters contained in \code{model}, e.g. "sj", and "name1, ..."
refers to variables in \code{data}, e.g. "Experiment".}

\item{error}{two-sided formula of the form
\code{par1+par2+... ~ name1+name2+...} where "par1, par2, ..." are
parameter containd in \code{error}, e.g. "sigma0" and "sigmaR", and
"name1, ..." refers to variables in \code{data}. If the same values
of "par1, ..." should be assumed for all data, "name1" can be "1".}

\item{log}{logical indicating whether all parameters are fitted on
log-scale.}

\item{normalize}{logical indicating whether the fixed effect
parameter should be normalized to unit mean.}

\item{verbose}{logical, print out information about each fit}

\item{normalize_input}{logical, if TRUE the input will be normalized before
scaling. see \link{splitData}.}
}
\value{
Object of class \code{aligned}, i.e. a data frame of the
alignment result containing an attribute "outputs":
 a list of data frames
\describe{
\item{prediction}{original data with value and sigma replaced by
                  the predicted values and sigmas}
\item{scaled}{original data with the values transformed according
              to the inverse model, i.e. \code{model} solved for
              the first parameter in \code{fixed}, e.g. "ys".
              Sigma values are computed by error propagation
              from the inverse model equation.}
\item{aligned}{the reduced data with the fixed effects and their
               uncertainty, only. The result of the alignment
               algorithm.}
\item{original}{the original data}
\item{parameter}{original data augmented by parameter columns.
                 Parameters in each row correspond to the levels of
                 fixed, latent or error as passed to alignME().
                 Used for initialization or parameter values when
                 refitting with modified model.}
}

The estimated parameters are returned by the attribute "parameters".
}
\description{
The function deals primarily with time-course data of different
targets which have been measured under different experimental
conditions and whose measured values might be on a different
scale, e.g. because of different amplification. The algorithm
determines the different scaling and estimates the time-course on
a common scale.
}
\details{
Alignment of time-course data is achieved by an alignment
model which explains the observed data by a function mixing
fixed effects, usually parameters reflecting the "underlying"
time-course, and latent variables, e.g. scaling parameters taking
account for effects like different amplification or loading, etc.
Depending on the measurement technique, the data has constant
relative error, or constant absolute error or even a combination
of those. This error is described by an error function. The error
parameters are usually global, i.e. the same parameter values are
assumed for all data points.
}
\examples{

}
\seealso{
\link{read.wide} to read data in a wide column format and
get it in the right format for \code{alignME()}.
}
