% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_functions.R
\name{analyze_blocks}
\alias{analyze_blocks}
\title{Method to analyze which elements of the given matrix with <number of unique
set of scaling parameters> columns, and <unique set of distinguish
parameters> rows, are on the same scale (same block) and pass a list of lists
of the respective indices.
Each row describes a unique set of distinguishable conditions, e.g. a
specific target measured under a specific condition at one time point.
The columns describe the sets of scaling parameters as target name and
gel (in case of western blot).}
\usage{
analyze_blocks(block_matrix)
}
\arguments{
\item{block_matrix}{Input matrix, with entries equal to 1 wherever the set of
distinguishable effects (row) is measured und the respective set of scaling
effects (column), i.e. each row has a one for each scaling under which it was
measured and each column has a one indicating which distinguishable effects
where measured at the respective scaling.
All entries are either one or zero.}
}
\value{
List with one entry per scale. Each entry contains a list of row
indices of the sets of distinguishable effects measured under the respective
scale.
}
\description{
Method to analyze which elements of the given matrix with <number of unique
set of scaling parameters> columns, and <unique set of distinguish
parameters> rows, are on the same scale (same block) and pass a list of lists
of the respective indices.
Each row describes a unique set of distinguishable conditions, e.g. a
specific target measured under a specific condition at one time point.
The columns describe the sets of scaling parameters as target name and
gel (in case of western blot).
}
