% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_exported_functions.R
\name{llr_test}
\alias{llr_test}
\title{Method for hypothesis testing}
\usage{
llr_test(H0, H1, check = TRUE)
}
\arguments{
\item{H0}{output of \link{align_me} obeying the null hypothesis. A special
case of \code{H1}.}

\item{H1}{output of \link{align_me}, the general case}
}
\value{
list with the log-likelihood ratio, statistical information and the
numerical p-value calculated by the evaluating the chi-squared distribution
at the present log-likelihood ratio with the current degrees of freedom.
}
\description{
Two outputs of \link{align_me} can be tested as nested hypothesis. This can
be used to test if e.g. buffer material influence can be neglected or a
specific measurement point is an outlier.
}
