% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{my_hello}
\alias{my_hello}
\title{blotIt out to list converter}
\usage{
my_hello()
}
\arguments{
\item{out}{'out' file as produced by \code{alignME()}.}

\item{use_factors}{Logical, if set to \code{TRUE} the name columns are converted to
factors. By this, the plotted order of target will be same order as they are
in the source data file fad through \link{read.wide} to \link{alignME}.}
}
\description{
Takes the output of blotIt2 and converts it to a list by reading out the
attributes.
}
