% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_functions.R
\name{objective_function}
\alias{objective_function}
\title{Objective function for trust region optimizer}
\usage{
objective_function(
  current_parameters,
  pass_parameter_list,
  pass_parameter_list2,
  calculate_derivative = TRUE
)
}
\description{
Objective function for trust region optimizer
}
