% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_profile_tools.R
\name{parframe}
\alias{parframe}
\title{Generate a parameter frame}
\usage{
parframe(
  x = NULL,
  parameters = colnames(x),
  metanames = NULL,
  obj.attributes = NULL
)
}
\arguments{
\item{x}{data.frame.}

\item{parameters}{character vector, the names of the parameter columns.}

\item{metanames}{character vector, the names of the meta-information columns.}

\item{obj.attributes}{character vector, the names of the objective function attributes.}
}
\value{
An object of class \code{parframe}, i.e. a data.frame with attributes for the
different names. Inherits from data.frame.
}
\description{
A parameter frame is a data.frame where the rows correspond to different
parameter specifications. The columns are divided into three parts. (1) the meta-information
columns (e.g. index, value, constraint, etc.), (2) the attributes of an objective function
(e.g. data contribution and prior contribution) and (3) the parameters.
}
\details{
Parameter frames can be subsetted either by \code{[ , ]} or by \code{subset}. If
\code{[ , index]} is used, the names of the removed columns will also be removed from
the corresponding attributes, i.e. metanames, obj.attributes and parameters.
}
\seealso{
\link{profile}, \link{mstrust}
}
