% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_exported_functions.R
\name{plot_align_me}
\alias{plot_align_me}
\title{All-in-one plot function for blotIt()}
\usage{
plot_align_me(
  out_list,
  ...,
  plot_points = "aligned",
  plot_line = "aligned",
  spline = FALSE,
  scales = "free",
  align_zeros = TRUE,
  plot_caption = TRUE,
  ncol = NULL,
  my_colors = NULL,
  duplicate_zero_points = FALSE,
  my_order = NULL
)
}
\arguments{
\item{out_list}{\code{out_list} file as produced by \link{align_me}, a list
of data.frames.}

\item{...}{Logical expression used for subsetting the data frames, e.g.
\code{name == "pERK1" & time < 60}.

\describe{
To reproduce the known function \code{plot1}, \code{plot2} and \code{plot3},
use:
\item{plot1}{
\code{plot_points} = 'original', \code{plot_line} = 'prediction'
}
\item{plot2}{
\code{plot_points} = 'scaled', \code{plot_line} = 'aligned'
}
\item{plot3}{
\code{plot_points} = 'aligned', \code{plot_line} = 'aligned'
}
}}

\item{plot_points}{String to specify which data set should be plotted in form
of points with corresponding error bars. It must be one of
\code{c("original", "scaled", "prediction", "aligned")}.}

\item{plot_line}{Same as above but with a line and error band.}

\item{spline}{Logical, if set to \code{TRUE}, what is specified as
\code{plot_line} will be plotted as a smooth spline instead of straight lines
between points.}

\item{scales}{String passed as \code{scales} argument to \link{facet_wrap}.}

\item{align_zeros}{Logical, if \code{TRUE}, the zero ticks are aligned
between the facets.}

\item{plot_caption}{Logical, if \code{TRUE}, a caption describing the plotted
data is added to the plot.}

\item{ncol}{Numerical passed as \code{ncol} argument to
\link{facet_wrap}.}

\item{my_colors}{list of custom color values as taken by the \code{values}
argument in the \link{scale_color_manual} method for \code{ggplot} objects.}

\item{duplicate_zero_points}{Logical, if set to \code{TRUE} all zero time
points are assumed to belong to the first condition. E.g. when the different
conditions consist of treatments added at time zero. Default is \code{FALSE}.}

\item{my_order}{Optional list of target names in the custom order that will
be used for faceting.}
}
\value{
ggplot object
}
\description{
Takes the output of \link{align_me} and generates graphs. Which data will be
plotted can then be specified separately.
}
\author{
Severin Bang and Svenja Kemmer
}
